ML Bài 4 — 'Hiểu kiểu dữ liệu trước khi train' (spec C1-L4, mở màn M2 Data Readiness) theo 3 quyết định user: hero ỐNG KÍNH DTYPE tương tác (renderDtypeLens — bấm thẻ cột lật dtype LƯU TRỮ vs NGHĨA THẬT + hành động modeling; đủ 4 thẻ bắt buộc → câu đố chốt completion-rule 'cột SỐ ĐẾM thật giữa 3 int64' = missed_classes, sai student_id → fb 'trung bình mã số vô nghĩa'), story FILE MỚI Ticket #04 (student_profile 200×6 độc lập bảng Bài 3 — tránh mâu thuẫn số), step 3 map 3 TRẠM = 3 VÒNG spec (VAI TRÒ roles_split → PHÂN NHÓM KIỂU type_groups bảng tô 4 màu ngữ nghĩa → ĐÓNG GÓI readiness card X(200×4)+encoding warning), 3 mồi bẫy (ID vào continuous · target lọt binary · 0/1 xếp nhóm đếm); ml_flow_map v5 thêm 2 result_kind type_groups+readiness dùng lại được; content chuẩn đợt5-6 (glossary 6 thẻ + 3 MCQ ôn glossary + minigame 6 cột→3 ngăn nghĩa + step4≠step3 thêm df.dtypes/unique); grade_lesson4 semantic có sẵn khớp (Risk bắt scholarship-vào-nhóm-số). Vá bug soi-screenshot: .dlens-reveal{display:flex} override hidden → lộ nghĩa khi chưa bấm → [hidden]{display:none}. ml_lab ép .astype(int64) missed/scholarship (dtype ổn định Win↔Pyodide). verify_b4 30/30 · 0 pageerror + regression B1/B2/B3 32/31/32 sạch; css v12·content v10·shell v19·flowmap v5

+++ static/py/ml_lab.py
@@ -322,9 +322,11 @@ def load_student_profile(shuffle_seed=None):
     rng = np.random.RandomState(1401)
     n = 200
     study_hours = np.round(rng.uniform(0.5, 10.0, n), 1)
-    missed_classes = rng.poisson(3, n).clip(0, 12)
+    # B4 đợt 7: ép int64 tường minh — poisson/astype(int) ra int32 trên Windows nhưng
+    # int64 trên Pyodide; bài giảng dạy "3 cột CÙNG int64" nên dtype phải ổn định mọi nền.
+    missed_classes = rng.poisson(3, n).clip(0, 12).astype(np.int64)
     major = np.array(['ICT', 'DS', 'Space'])[rng.randint(0, 3, n)]
-    scholarship = (rng.uniform(0, 1, n) < 0.3).astype(int)
+    scholarship = (rng.uniform(0, 1, n) < 0.3).astype(np.int64)
     score = 6.0 * study_hours - 1.5 * missed_classes + 3.0 * scholarship + rng.normal(0, 8, n) + 25
     pass_fail = (score >= 50).astype(int)
     df = pd.DataFrame({
